feat(preview): add mocked RestaurantScreen preview using fake ViewModel and mock Restaurant data. Closes #14

+++ app/src/main/java/com/sadaquekhan/justeatassessment/domain/model/RestaurantResponse.kt
@@ -7,7 +7,7 @@ data class RestaurantResponse(
 data class RestaurantDto(
     val id: String,
     val name: String,
-    val rating: Float,
+    val rating: Double,
     val cuisineType: String,
     val address: String
 ) {
